% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_controls.R
\name{check_missing_LGA_names}
\alias{check_missing_LGA_names}
\title{check_missing_LGA_names}
\usage{
check_missing_LGA_names(data, control)
}
\arguments{
\item{data}{combined_dataset_no_controls.xlsx}

\item{control}{one of the control datasets}
}
\value{
NULL or vector of names that are not directly named in the control dataset. These names need to be mapped manually.
}
\description{
check_missing_LGA_names
}
