% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_demographic_statistics.R
\name{clean_demographic_statistics}
\alias{clean_demographic_statistics}
\title{clean_demographic_statistics}
\usage{
clean_demographic_statistics(raw_sheet_path)
}
\arguments{
\item{raw_sheet_path}{Path to age and sex demographics.xlsx.}
}
\value{
A table with \% of men, \% of females, total males and total females by state and LGA by year
}
\description{
clean_demographic_statistics
}
