% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_demographic_statistics.R
\name{clean_demographic_table}
\alias{clean_demographic_table}
\title{clean_demographic_table}
\usage{
clean_demographic_table(raw_data)
}
\arguments{
\item{raw_data}{Data loaded from a table in age and sex demographics.xlsx.}
}
\value{
Table which lists \% of population as sex for age buckets and total population for each state and LGA by year
}
\description{
clean_demographic_table
}
