% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_nsw_datasets.R
\name{clean_nsw_sheet}
\alias{clean_nsw_sheet}
\title{clean_nsw_sheet}
\usage{
clean_nsw_sheet(raw_sheet, file_name)
}
\arguments{
\item{raw_sheet}{A dataframe. The raw sheet to clean}

\item{file_name}{The name of the file. Used to retrieve the month and year}
}
\value{
A dataframe holding the cleaned LGA sheet
}
\description{
This cleans each individual NSW quarterly rent excel LGA sheet.
}
