% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_personal_income_statistics.R
\name{clean_personal_income_statistics}
\alias{clean_personal_income_statistics}
\title{clean_personal_income_statistics}
\usage{
clean_personal_income_statistics(file_path)
}
\arguments{
\item{file_path}{File path to income,earners by geography.xlsx}
}
\value{
Cleaned Table 1.5 of the file, i.e. earner demographics split by LGA
}
\description{
clean_personal_income_statistics
}
