% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sa_datasets.R
\name{clean_pre_2020_Q3_sa_sheet}
\alias{clean_pre_2020_Q3_sa_sheet}
\title{clean_pre_2020_Q3_sa_sheet}
\usage{
clean_pre_2020_Q3_sa_sheet(raw_sheet, file_name)

clean_pre_2020_Q3_sa_sheet(raw_sheet, file_name)
}
\arguments{
\item{raw_sheet}{A dataframe. The raw sheet to clean}

\item{file_name}{The name of the file. Used to retrieve the month and year}
}
\value{
A dataframe holding the cleaned LGA sheet

A dataframe holding the cleaned LGA sheet
}
\description{
This cleans each individual SA quarterly rent excel LGA sheet before 2020 Quarter 3 (inclusive).

This cleans each individual SA quarterly rent excel LGA sheet post 2020 Quarter 3 (exclusive).
}
