% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_qld_dataset.R
\name{clean_qld_dataset}
\alias{clean_qld_dataset}
\title{clean_qld_dataset}
\usage{
clean_qld_dataset(raw_sheet_path)
}
\arguments{
\item{raw_sheet_path}{File path to raw qld data file}
}
\value{
A dataframe with all the data from the excel sheet
}
\description{
This function takes the raw qld dataset (that has had some manual modifications made to it) and appends
all the sheets together into the required shape
}
