% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_qld_dataset.R
\name{clean_qld_sheet}
\alias{clean_qld_sheet}
\title{clean_qld_sheet}
\usage{
clean_qld_sheet(raw_sheet, sheet_name)
}
\arguments{
\item{raw_sheet}{A dataframe passed from clean_vic_dataset}

\item{sheet_name}{A string passed from clean_vic_dataset}
}
\value{
A dataframe containing the cleaned sheet
}
\description{
This function takes either qld-rents or qld-new-bonds from the qld data file, changes its shape, cleans values and columns and adds
additional columns
}
