% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sa_datasets.R
\name{clean_sa_datasets}
\alias{clean_sa_datasets}
\title{clean_sa_datasets}
\usage{
clean_sa_datasets(folder_path)
}
\arguments{
\item{folder_path}{A string. Path to where you are holding the raw data files.}
}
\value{
A dataframe. Final cleaned dataframe containing cleaned data from all data provided in the folder path.
}
\description{
This cleans all the datasets retrieved from download_sa_datasets and combines it into one dataframe.
}
\seealso{
\code{\link[=download_sa_datasets]{download_sa_datasets()}}
}
