% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vic_dataset.R
\name{clean_vic_dataset}
\alias{clean_vic_dataset}
\title{clean_vic_dataset}
\usage{
clean_vic_dataset(raw_sheet_path)
}
\arguments{
\item{raw_sheet_path}{File path to raw vic data file}
}
\value{
A dataframe with all the data from the excel sheet
}
\description{
This function takes the raw vic dataset (that has had some manual modifications made to it) and appends
all the sheets together into the required shape
}
