% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vic_dataset.R
\name{clean_vic_sheet}
\alias{clean_vic_sheet}
\title{clean_vic_sheet}
\usage{
clean_vic_sheet(raw_sheet, sheet_name)
}
\arguments{
\item{raw_sheet}{A dataframe passed from clean_vic_dataset}

\item{sheet_name}{A string passed from clean_vic_dataset}
}
\value{
A dataframe containing the cleaned sheet
}
\description{
This function takes a single sheet from the vic data file, changes its shape, cleans values and columns and adds
additional columns
}
