% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_demographic_statistics.R
\name{convert_demographic_stats_to_percentage}
\alias{convert_demographic_stats_to_percentage}
\title{convert_demographic_stats_to_percentage}
\usage{
convert_demographic_stats_to_percentage(data, age_groups)
}
\arguments{
\item{data}{data to calculate \% from}

\item{age_groups}{column names of age groups to calculate as a \% of row total}
}
\value{
Resident population table in \% of row total
}
\description{
convert_demographic_stats_to_percentage
}
