% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rent_and_new_bonds.R
\name{convert_text_to_ggplot}
\alias{convert_text_to_ggplot}
\title{convert_text_to_ggplot}
\usage{
convert_text_to_ggplot(text)
}
\arguments{
\item{text}{A string. Text to create as its own plot.}
}
\value{
A text ggplot object containing only the text with no axes or background.
}
\description{
A small helper that draws a bit of text as its own plot, allowing the text
to be used with the rest of ggplot and patchwork function. This is helpful for
arranging free standing text.
}
