% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nsw_datasets.R
\name{create_nsw_file_name}
\alias{create_nsw_file_name}
\title{create_nsw_file_name}
\usage{
create_nsw_file_name(link)
}
\arguments{
\item{link}{url to convert to a valid file name}
}
\value{
A valid file name in the form "NSW month year quarter"
}
\description{
create_nsw_file_name
}
