% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sa_datasets.R
\name{download_sa_datasets}
\alias{download_sa_datasets}
\title{download_sa_datasets}
\usage{
download_sa_datasets(file_path)
}
\arguments{
\item{file_path}{Location to save files}
}
\value{
Excel files saved in file path
}
\description{
This function takes data from the SA gov website and downloads all rent tables from 2018 onwards
}
