% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_controls.R
\name{join_controls}
\alias{join_controls}
\title{join_controls}
\usage{
join_controls(data, controls)
}
\arguments{
\item{data}{combined_dataset_no_controls.xlsx}

\item{controls}{cleaned income and demographic data}
}
\value{
data with controls join by State, LGA and Year.
}
\description{
join_controls
}
\seealso{
\code{\link[=clean_personal_income_statistics]{clean_personal_income_statistics()}}, \code{\link[=clean_demographic_statistics]{clean_demographic_statistics()}}
}
