% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rent_and_new_bonds.R
\name{plot_bond}
\alias{plot_bond}
\title{plot_bonds}
\usage{
plot_bond(data, palette)
}
\arguments{
\item{data}{Data to plot.}

\item{palette}{Named vector. Maps each state to a color.}
}
\value{
New bonds lodged plotted by quarter start date and grouped by state.
}
\description{
plot_bonds
}
