% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_nsw_datasets.R
\name{clean_nsw_datasets}
\alias{clean_nsw_datasets}
\title{clean_nsw_datasets}
\usage{
clean_nsw_datasets(folder_path)
}
\arguments{
\item{folder_path}{A string. Path to where you are holding the raw data files.}
}
\value{
A dataframe. Final cleaned dataframe containing cleaned data from all data provided in the folder path.
}
\description{
This cleans all the datasets retrieved from download_nsw_datasets and combines it into one dataframe.
}
\seealso{
\code{\link[=download_nsw_datasets]{download_nsw_datasets()}}
}
