% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_keys.R
\name{create_demographic_table_key}
\alias{create_demographic_table_key}
\title{create_demographic_table_key}
\usage{
create_demographic_table_key()
}
\value{
A tibble mapping the missing names in combined_dataset_no_controls.xlsx to their equivalent names in age and sex demographics.xlsx
}
\description{
Some LGA names in combined_dataset_no_controls.xlsx won't be present in age and sex demographics.xlsx.
This just due to different names for the same LGA. This function lists the names in combined_dataset_no_controls.xlsx,
that are not present in age and sex demographics.xlsx.
}
\details{
It then lists, the equivalent names in age and sex demographics.xlsx in the vector matching_LGA_demo_names
}
