% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_keys.R
\name{create_keys}
\alias{create_keys}
\title{create_keys}
\usage{
create_keys()
}
\value{
A list of data-frames mapping missing LGA names from the rent dataset to the LGA names in the control tables
}
\description{
creates a mapping key for income earners,by geography Table 1.5 LGA names to LGA names in combined_dataset_no_controls.xlsx.
This only maps names that are not identical. The key will be used to replace the values before a left join occurs.
}
