% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rent_and_new_bonds.R
\name{plot_rent}
\alias{plot_rent}
\title{plot_rent}
\usage{
plot_rent(data, palette)
}
\arguments{
\item{data}{Data to plot.}

\item{palette}{Named vector. Maps each state to a color.}
}
\value{
Median weekly rent plotted by quarter start date and grouped by state.
}
\description{
plot_rent
}
