% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rent_and_new_bonds.R
\name{plot_rent_bonds_segment}
\alias{plot_rent_bonds_segment}
\title{plot_rent_bonds_segment}
\usage{
plot_rent_bonds_segment(data, treated, subset_details, palette)
}
\arguments{
\item{data}{A data frame. Contains the data to subset.}

\item{treated}{A string. The treated state to plot.}

\item{subset_details}{A dataframe. Contains the details to filter the data by. Should have the columns:
\enumerate{
\item Control. A string. The control state to plot.
\item Dwelling. A string. The dwelling type to subset the data by.
\item Bedroom. A string. The number of bedrooms to subset the data by.
}}

\item{palette}{Named vector. Maps each state to a color.}
}
\value{
Plots of median weekly rent and control by dwelling and bedroom with the treated and control states on the same plot.
}
\description{
For each row of \code{subset_details}, plots median weekly rent and new bonds for the
treated state against that row's control state.
}
